Add tests for C++ select_points routine.
The behavior of the C function was fixed in f5fe2ce, but the C++
function still adds a point even when a maximum was already
enclosed. These tests will help ensure identical behavior between the
two functions.

(Ideally, one of the functions would just call the other so that this
wasn't necessary.)