Decrease max_t in simulations.
In bracket_maximum(), if the curve appears to be monotonically
increasing in every iteration, the range eventually returned will be
small and close to max_t. The guess used in estimate_derivatives() is
in the middle of this range. If it's close enough to max_t, the
stencil points to the right of the guess may exceed max_t and result
in calling the log-likelihood function with a value that is out of
bounds.

This may also occur with min_t when the curve appears to be
monotonically decreasing, but I haven't seen it happen.